define interfaces where they are needed
